# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Flier API",
    "description": "RESTful API for Flier - Professional Email Campaign Management Platform. This API provides endpoints for managing contacts, campaigns, templates, and email sending through Gmail API integration.",
    "version": "1.0.0",
    "contact": {
      "name": "Flier Support",
      "url": "https://github.com/Aditya190803/flier"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://sendflier.tech",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/send-email": {
      "post": {
        "tags": ["Email"],
        "summary": "Send email campaign",
        "description": "Send personalized emails to multiple recipients",
        "operationId": "sendEmailCampaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Emails sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendEmailResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/gdpr/export": {
      "get": {
        "tags": ["GDPR"],
        "summary": "Export user data",
        "description": "Download all user data as JSON (GDPR data portability)",
        "operationId": "exportUserData",
        "responses": {
          "200": {
            "description": "JSON file containing all user data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GDPRExport"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SendEmailRequest": {
        "type": "object",
        "oneOf": [
          {
            "required": ["personalizedEmails"],
            "properties": {
              "personalizedEmails": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "to": { "type": "string", "format": "email" },
                    "subject": { "type": "string" },
                    "message": { "type": "string" },
                    "originalRowData": { "type": "object" },
                    "attachments": { "type": "array" }
                  }
                }
              }
            }
          },
          {
            "required": ["recipients", "subject", "content"],
            "properties": {
              "recipients": {
                "type": "array",
                "items": { "type": "string", "format": "email" }
              },
              "subject": { "type": "string" },
              "content": { "type": "string" },
              "variants": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "subject": { "type": "string" },
                    "content": { "type": "string" }
                  }
                }
              }
            }
          }
        ],
        "properties": {
          "campaignId": { "type": "string" },
          "trackingEnabled": { "type": "boolean", "default": true },
          "verifyBeforeSending": { "type": "boolean", "default": false }
        }
      },
      "SendEmailResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": ["success", "error"]
                },
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "summary": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "sent": {
                "type": "integer"
              },
              "failed": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "GDPRExport": {
        "type": "object",
        "properties": {
          "exportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object"
          },
          "contacts": {
            "type": "array"
          },
          "campaigns": {
            "type": "array"
          },
          "templates": {
            "type": "array"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Unauthorized"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth.js session cookie"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}